+++ gradio/utils.py
@@ -41,6 +41,9 @@ import fsspec.asyn
 import httpx
 import matplotlib.pyplot as plt
 import requests
+from markdown_it import MarkdownIt
+from mdit_py_plugins.dollarmath.index import dollarmath_plugin
+from mdit_py_plugins.footnote.index import footnote_plugin
 from pydantic import BaseModel, Json, parse_obj_as
 
 import gradio
@@ -893,3 +896,28 @@ def abspath(path: str | Path) -> Path:
         return Path.cwd() / path
     else:
         return Path(path).resolve()
+
+
+def get_markdown_parser() -> MarkdownIt:
+    md = (
+        MarkdownIt(
+            "js-default",
+            {
+                "linkify": True,
+                "typographer": True,
+                "html": True,
+            },
+        )
+        .use(dollarmath_plugin, renderer=tex2svg, allow_digits=False)
+        .use(footnote_plugin)
+        .enable("table")
+    )
+
+    # Add target="_blank" to all links. Taken from MarkdownIt docs: https://github.com/executablebooks/markdown-it-py/blob/master/docs/architecture.md
+    def render_blank_link(self, tokens, idx, options, env):
+        tokens[idx].attrSet("target", "_blank")
+        return self.renderToken(tokens, idx, options, env)
+
+    md.add_render_rule("link_open", render_blank_link)
+
+    return md
